Fix jet1090 altitude type and clippy warnings
Update jet1090 crate to use i32 for altitude fields to match the
rs1090 library changes. This ensures type consistency across the
workspace and resolves clippy errors.

Changes:
- snapshot.rs: Change Snapshot.altitude from Option<u16> to Option<i32>
- table.rs: Fix altitude comparison to use i32::abs_diff instead of u16
- bds05.rs: Remove useless i32::from conversion in test (clippy warning)

All tests pass and clippy --workspace --all-targets passes cleanly.

+++ crates/jet1090/src/snapshot.rs
@@ -36,7 +36,9 @@ pub struct Snapshot {
     /// WGS84 longitude angle in degrees
     pub longitude: Option<f64>,
     /// Barometric altitude in feet, expressed in ISA
-    pub altitude: Option<u16>,
+    /// Can be negative for airports below sea level (e.g., Amsterdam Schiphol)
+    /// Can be positive up to ~50,000 ft for cruise altitude
+    pub altitude: Option<i32>,
     /// Altitude selected in the FMS
     pub selected_altitude: Option<u16>,
     /// Ground speed, in knots
